More packages

+++ packages/toggleads/uninstall.sh
@@ -2,6 +2,6 @@ set -eu
 
 SCRIPTLET=/mnt/us/documents/toggle-ads.sh
 
-if [ -f "${SCRIPTLET}" ] && cmp -s scriptlets/toggle-ads.sh "${SCRIPTLET}"; then
+if [ -f "${SCRIPTLET}" ] && [ ! -L "${SCRIPTLET}" ] && cmp -s scriptlets/toggle-ads.sh "${SCRIPTLET}"; then
   rm -f "${SCRIPTLET}"
 fi
