+++ packages/updateblockstatus/uninstall.sh
@@ -2,6 +2,6 @@ set -eu
 
 SCRIPTLET=/mnt/us/documents/updateblock.sh
 
-if [ -f "${SCRIPTLET}" ] && cmp -s scriptlets/updateblock.sh "${SCRIPTLET}"; then
+if [ -f "${SCRIPTLET}" ] && [ ! -L "${SCRIPTLET}" ] && cmp -s scriptlets/updateblock.sh "${SCRIPTLET}"; then
   rm -f "${SCRIPTLET}"
 fi
